Parsing via TMX is in the loadFromTMX in ALevel, but ArrowTraps are broken with it? Someone might need to take a look at it. Will need to convert rest of levels at a later time.

+++ core/src/com/totlc/Actors/enemies/Spider.java
@@ -43,7 +43,7 @@ public class Spider extends AEnemy {
     private TextureAtlas walkTextureAtlas;
     private Animation<TextureRegion> walkAnimation;
 
-    public Spider(AssetManager assetManager, int x, int y, AMovement movement){
+    public Spider(AssetManager assetManager, float x, float y, AMovement movement){
         super(assetManager, new Rectangle(x, y, width, height), movement, hp, atk);
 
         //TODO: Remove after fixing hitboxes
